requirements.txt
# This file is expected by Heroku.
django-countries==6.1.2

-r requirements/production.txt
